move e2e tests into separate test suite
